% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{bindr_flxbl}
\alias{bindr_flxbl}
\title{Bind rows from dataframes in list}
\usage{
bindr_flxbl(dfs_list)
}
\arguments{
\item{dfs_list}{list of dataframes}
}
\value{
a dataframe
}
\description{
if dataframes have different column length, return them clustered by clength (in a list)
}
\examples{
bindr_flxbl(list(mtcars,mtcars))
}
