% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays.R
\name{dilute}
\alias{dilute}
\title{Explode dataframe to list of lists}
\usage{
dilute(..., row = F)
}
\arguments{
\item{...}{list of dataframes to merge together}

\item{row}{if TRUE, the output list will consist of the rows of the original dataframes as sublist-elements, otherwise the original dataframes columns (default: FALSE)}
}
\value{
a list containing row/column of all the dataframes in \code{...} as sublists
}
\description{
merges dataframes columns/rows from a list into a new list whereby every element is a list containing the column/row elements for each of the dataframes. All of the dataframes are be ultimately merged together in the same list.
}
\examples{
dilute(mtcars,iris)
dilute(mtcars,iris,row=T)
}
