% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays.R
\name{easin}
\alias{easin}
\title{Decompose a string into a vector}
\usage{
easin(str, sep = " ")
}
\arguments{
\item{str}{Input character vector}

\item{sep}{Delimiter to be used to subset \link{str} into chunks}
}
\value{
A vector of characters
}
\description{
Create a vector of strings from a string based on a delimiter
}
\examples{
fruits <- easin("apple berry banana pear")
}
