% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays.R
\name{highest_in}
\alias{highest_in}
\title{Find the data frame with greatest amount of columns named a certain way}
\usage{
highest_in(listOrEnv, ..., FUN = dplyr::starts_with, row = F)
}
\arguments{
\item{listOrEnv}{container of the input dataframes}

\item{...}{one or more attributes to search by, the type depends on the \link{FUN} accepted input (default:character)}

\item{FUN}{the function to apply for matching column or row names (default: \link[dplyr]{starts_with})}

\item{row}{if TRUE row names are used for search and scoring}
}
\value{
a dataframe from \link{listOrEnv}
}
\description{
Given a list/environment of data frames and one or more attributes, it finds the one dataframe having
most columns/rows named as specified by the attributes. The default function used for matching is \link[dplyr]{starts_with} (dplyr v.1.0.7)
}
\examples{
df1<-mtcars[1:9,];df2<-mtcars[10:16,]
highest_in(list(df1,df2),"Merc",row=T)
highest_in(list(df1,df2),c("Merc","Cadillac"),row=T)
highest_in(list(df1,df2),c("Merc","Valiant"),row=T)
}
