% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays.R
\name{lfus}
\alias{lfus}
\title{List fusion by name}
\usage{
lfus(list1, list2, labeling, follow = list1)
}
\arguments{
\item{list1}{list to be fused by name}

\item{list2}{list to be fused by name}

\item{labeling}{must be a character vector of length 2}

\item{follow}{specify which order should the resulting list have. If \code{list2}, the order will be the same of the second input list (default: \code{list1})}
}
\value{
a list containing all the elements of \code{list1} and \code{list2}
}
\description{
#fuses two lists by names, creating a new list where each sub-list contain the elements
of the original lists in the same ordered and labels as per \code{labeling}.
}
\examples{
la<-list(type="sunflower",available=F,quantity=1)
lb<-list(available=T,type="rosemary",quantity=3)
lfus(la,lb,c("A","B"))
}
